Initial work on an Alpine Linux APKBUILD
The goal here is to have an APKBUILD that outputs a package containing a
copy of the Rust compiler that will run on Alpine Linux (i.e. the
compiler is compiled using musl, and also has both musl and regular GNU
targets).  This should result in a statically-linked compiler that is
capable of outputing statically-linked binaries (or regular Linux
binaries).